vagrant: Avoid download of partial cached files
During VM provisioning, we cache some downloaded files outside the VM,
to speed up subsequent provisionings. If a cached file is present we
consider that it's safe to reuse. That is however not the case when the
download was interrupted (e.g., because of a network error or because VM
provisioning was interrupted). Ideally we would check the signature of
downloaded files, but the associated projects have different ways of
exposing the signatures.

This commit changes download_to() to download files in /tmp and move
them to the cache only once the download is succesfull. This behavior
requires the use of 'set -e'. The cache may still end with partial files
if provisioning is interrupted while the file is moved out of /tmp. That
operation is however much faster than the download and an interruption
therefore less likely.

Signed-off-by: Paul Chaignon <paul@cilium.io>